Update README.md
I developed a space invaders-style game infused with Star Wars themes. This game uses text boxes to display the player's score, and two functional buttons for navigation between three screens. It incorporates an array of sound effects and visual enhancements that respond to the player's actions, using logical and mathematical parameters for optimal functionality. 

One key aspect of the app's functionality is the disappearance and reappearance of the enemy ship. The ship becomes invisible upon being hit by a projectile, otherwise the ship would be permanently visible. I have included sound effects and visual sprites when the enemy ship is hit which enriches the user experience. To maintain consistency, the explosion sprite precisely covers the area impacted by the projectile rather than appearing randomly. After destroying the enemy ship, it no longer reappears in the same position. I implemented mathematical functions to generate a random integer that fits within the screen size, which makes the enemy ship spawn into a new location each time it is hit. 

The text box at the bottom of the display screen records the player's score and is incremented by each hit on the enemy ship. This scoring system gives the player a sense of achievement. Pressing the "End" button resets the player's score to zero and transitions them to the End Game Screen. 

I encountered challenges during the development process. Some images did not align with the screen or game theme. Additionally, there was an issue with screen navigation where the main screen often took precedence over the "welcome" screen. This can be easily remedied by opening the companion app while the “Welcome Screen” is open on the host device. Despite these issues, I consider this project a success and hope to expand upon the knowledge and skills I acquired in future.